Fix tests

+++ r1cs-std-new/src/bits/uint.rs
@@ -148,7 +148,7 @@ macro_rules! make_uint {
 
                     // Compute the maximum value of the sum so we allocate enough bits for
                     // the result
-                    let mut max_value = (operands.len() as u64) * u64::from($native::max_value());
+                    let mut max_value = (operands.len() as u128) * u128::from($native::max_value());
 
                     // Keep track of the resulting value
                     let mut result_value = Some(0u128);
